requirements.txt
aniso8601==4.0.1
Click==7.0
EditorConfig==0.12.2
Flask==1.0.2
flask-restplus==0.12.1
#GDAL==2.4.0
itsdangerous==1.1.0
Jinja2==2.10
jsbeautifier==1.8.9
jsonschema==2.6.0
MarkupSafe==1.1.0
pytz==2018.7
six==1.12.0
Werkzeug==0.14.1
Requests==2.23.0
boto3==1.13.24
xmltodict==0.12.0
PyPDF2==1.26.0
openpyxl==3.0.5
xlrd==1.2.0
python-docx==0.8.10
textract==1.6.3
